fonoapi_mock/src/test/java/an/y/ledov/mock/rest/TokenControllerTest.java
package an.y.ledov.mock.rest;

import an.y.ledov.mock.rest.dto.RestToken;
import com.fasterxml.jackson.databind.ObjectMapper;
import lombok.SneakyThrows;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;

import static org.assertj.core.api.Assertions.assertThat;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@WebMvcTest(controllers = {
    TokenController.class})
public class TokenControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @Test
    @SneakyThrows
    @DisplayName("should successfully call generate token endpoint")
    void shouldSuccessfullyCallGenerateTokenEndpoint() {
        var requestUrl = "/token/generate";

        var result = mockMvc.perform(get(requestUrl)
                .contentType(MediaType.APPLICATION_JSON))
            .andDo(print())
            .andExpect(status().isOk())
            .andExpect(content().contentType(MediaType.APPLICATION_JSON))
            .andReturn();

        var token = new ObjectMapper().readValue(
            result.getResponse().getContentAsString(),
            RestToken.class);

        assertThat(token).isNotNull();
    }

}
